Plugins: Move interface implementations into cpp file
