+++ plugins/baserobots/baserobots.cpp
@@ -15,6 +15,53 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "baserobots.h"
+//#include "plugindefs.h"
+#include <openrave/openrave.h> // should be included first in order to get boost throwing openrave exceptions
+
+OpenRAVE::RobotBasePtr CreateCollisionMapRobot(OpenRAVE::EnvironmentBasePtr penv, std::istream& sinput);
+OpenRAVE::RobotBasePtr CreateConveyorRobot(OpenRAVE::EnvironmentBasePtr penv, std::istream& sinput);
+
+void RegisterCollisionMapRobotReaders(std::list< OpenRAVE::UserDataPtr >& listRegisteredReaders);
+void RegisterConveyorReaders(std::list< OpenRAVE::UserDataPtr >& listRegisteredReaders);
+
+const std::string BaseRobotsPlugin::_pluginname = "BaseRobotsPlugin";
+
+BaseRobotsPlugin::BaseRobotsPlugin()
+{
+    RegisterCollisionMapRobotReaders(s_listRegisteredReaders);
+    RegisterConveyorReaders(s_listRegisteredReaders);
+    _interfaces[OpenRAVE::PT_Robot].push_back("CollisionMapRobot");
+    _interfaces[OpenRAVE::PT_Robot].push_back("Conveyor");
+}
+
+BaseRobotsPlugin::~BaseRobotsPlugin() {}
+
+OpenRAVE::InterfaceBasePtr BaseRobotsPlugin::CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv)
+{
+    switch(type) {
+    case OpenRAVE::PT_Robot:
+        if( interfacename == "collisionmaprobot") {
+            return CreateCollisionMapRobot(penv,sinput);
+        }
+        else if( interfacename == "conveyor" ) {
+            return CreateConveyorRobot(penv,sinput);
+        }
+        break;
+    default:
+        break;
+    }
+    return OpenRAVE::InterfaceBasePtr();
+}
+
+const RavePlugin::InterfaceMap& BaseRobotsPlugin::GetInterfaces() const
+{
+    return _interfaces;
+}
+
+const std::string& BaseRobotsPlugin::GetPluginName() const
+{
+    return _pluginname;
+}
 
 OPENRAVE_PLUGIN_API RavePlugin* CreatePlugin() {
     return new BaseRobotsPlugin();
