+++ plugins/baserobots/baserobots.h
@@ -17,55 +17,18 @@
 #ifndef OPENRAVE_PLUGIN_BASEROBOTS_H
 #define OPENRAVE_PLUGIN_BASEROBOTS_H
 
-#include "plugindefs.h"
 #include <openrave/plugin.h>
 
-OpenRAVE::RobotBasePtr CreateCollisionMapRobot(OpenRAVE::EnvironmentBasePtr penv, std::istream& sinput);
-OpenRAVE::RobotBasePtr CreateConveyorRobot(OpenRAVE::EnvironmentBasePtr penv, std::istream& sinput);
-
-void RegisterCollisionMapRobotReaders(std::list< OpenRAVE::UserDataPtr >& listRegisteredReaders);
-void RegisterConveyorReaders(std::list< OpenRAVE::UserDataPtr >& listRegisteredReaders);
-
 struct BaseRobotsPlugin : public RavePlugin {
-    BaseRobotsPlugin()
-    {
-        RegisterCollisionMapRobotReaders(s_listRegisteredReaders);
-        RegisterConveyorReaders(s_listRegisteredReaders);
-        _interfaces[PT_Robot].push_back("CollisionMapRobot");
-        _interfaces[PT_Robot].push_back("Conveyor");
-    }
+    BaseRobotsPlugin();
+    ~BaseRobotsPlugin() override;
 
-    ~BaseRobotsPlugin() override {}
-
-    OpenRAVE::InterfaceBasePtr CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv) override
-    {
-        switch(type) {
-        case PT_Robot:
-            if( interfacename == "collisionmaprobot") {
-                return CreateCollisionMapRobot(penv,sinput);
-            }
-            else if( interfacename == "conveyor" ) {
-                return CreateConveyorRobot(penv,sinput);
-            }
-            break;
-        default:
-            break;
-        }
-        return OpenRAVE::InterfaceBasePtr();
-    }
-
-    const InterfaceMap& GetInterfaces() const override
-    {
-        return _interfaces;
-    }
-
-    const std::string& GetPluginName() const override
-    {
-        static std::string pluginname = "BaseRobotsPlugin";
-        return pluginname;
-    }
+    OpenRAVE::InterfaceBasePtr CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv) override;
+    const InterfaceMap& GetInterfaces() const override;
+    const std::string& GetPluginName() const override;
 
 private:
+    static const std::string _pluginname;
     InterfaceMap _interfaces;
     std::list<OpenRAVE::UserDataPtr> s_listRegisteredReaders;
 };
