+++ plugins/fclrave/fclrave.cpp
@@ -14,6 +14,35 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "fclrave.h"
+#include "plugindefs.h"
+#include "fclcollision.h"
+
+const std::string FCLRavePlugin::_pluginname = "FCLRavePlugin";
+
+FCLRavePlugin::FCLRavePlugin()
+{
+    _interfaces[OpenRAVE::PT_CollisionChecker].push_back("fcl_");
+}
+
+FCLRavePlugin::~FCLRavePlugin() {}
+
+OpenRAVE::InterfaceBasePtr FCLRavePlugin::CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv)
+{
+    if( type == OpenRAVE::PT_CollisionChecker && interfacename == "fcl_" ) {
+        return boost::make_shared<fclrave::FCLCollisionChecker>(penv, sinput);
+    }
+    return OpenRAVE::InterfaceBasePtr();
+}
+
+const RavePlugin::InterfaceMap& FCLRavePlugin::GetInterfaces() const
+{
+    return _interfaces;
+}
+
+const std::string& FCLRavePlugin::GetPluginName() const
+{
+    return _pluginname;
+}
 
 OPENRAVE_PLUGIN_API RavePlugin* CreatePlugin() {
     return new FCLRavePlugin();
