+++ plugins/mobyrave/mobyrave.h
@@ -15,64 +15,18 @@
 #ifndef OPENRAVE_PLUGIN_MOBYRAVE_H
 #define OPENRAVE_PLUGIN_MOBYRAVE_H
 
-#include "mobyphysics.h"
-#include "mobycontroller.h"
-#include "mobyreplaycontroller.h"
-
-// create moby physics shared pointer
-OpenRAVE::PhysicsEngineBasePtr CreateMobyPhysics(OpenRAVE::EnvironmentBasePtr penv, std::istream& sinput)
-{
-    return boost::make_shared<MobyPhysics>(penv, sinput);
-}
+#include <openrave/plugin.h>
 
 struct MobyRavePlugin : public RavePlugin {
-    MobyRavePlugin()
-    {
-        s_listRegisteredReaders.push_back(RaveRegisterXMLReader(OpenRAVE::PT_PhysicsEngine,"mobyphysics",MobyPhysics::CreateXMLReader));
-        s_listRegisteredReaders.push_back(RaveRegisterXMLReader(OpenRAVE::PT_Controller,"mobycontroller",MobyController::CreateXMLReader));
-        s_listRegisteredReaders.push_back(RaveRegisterXMLReader(OpenRAVE::PT_Controller,"mobyreplaycontroller",MobyReplayController::CreateXMLReader));
-        _interfaces[PT_PhysicsEngine].push_back("moby");
-        _interfaces[OpenRAVE::PT_Controller].push_back("moby");
-        _interfaces[OpenRAVE::PT_Controller].push_back("mobyreplay");
-    }
+    MobyRavePlugin();
+    ~MobyRavePlugin() override;
 
-    ~MobyRavePlugin() override {}
-
-    OpenRAVE::InterfaceBasePtr CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv) override
-    {
-        switch( type ) {
-        case PT_PhysicsEngine:
-            if( interfacename == "moby" )
-            {
-                return CreateMobyPhysics(penv,sinput);
-            }
-            break;
-        case OpenRAVE::PT_Controller:
-            if( interfacename == "moby") {
-                return boost::make_shared<MobyController>(penv,sinput);
-            }
-            else if( interfacename == "mobyreplay") {
-                return boost::make_shared<MobyReplayController>(penv,sinput);
-            }
-            break;
-        default:
-            break;
-        }
-        return OpenRAVE::InterfaceBasePtr();
-    }
-
-    const InterfaceMap& GetInterfaces() const override
-    {
-        return _interfaces;
-    }
-
-    const std::string& GetPluginName() const override
-    {
-        static std::string pluginname = "MobyRavePlugin";
-        return pluginname;
-    }
+    OpenRAVE::InterfaceBasePtr CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv) override;
+    const InterfaceMap& GetInterfaces() const override;
+    const std::string& GetPluginName() const override;
 
 private:
+    static const std::string _pluginname;
     InterfaceMap _interfaces;
     std::list<OpenRAVE::UserDataPtr> s_listRegisteredReaders;
 };
