+++ plugins/rmanipulation/rmanipulation.cpp
@@ -14,6 +14,56 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "rmanipulation.h"
+#include "plugindefs.h"
+
+OpenRAVE::ModuleBasePtr CreateBaseManipulation(OpenRAVE::EnvironmentBasePtr penv);
+//OpenRAVE::ModuleBasePtr CreateTaskCaging(OpenRAVE::EnvironmentBasePtr penv);
+OpenRAVE::ModuleBasePtr CreateTaskManipulation(OpenRAVE::EnvironmentBasePtr penv);
+OpenRAVE::ModuleBasePtr CreateVisualFeedback(OpenRAVE::EnvironmentBasePtr penv);
+
+RManipulationPlugin::RManipulationPlugin()
+{
+    _interfaces[PT_Module].push_back("BaseManipulation");
+    _interfaces[PT_Module].push_back("TaskManipulation");
+    _interfaces[PT_Module].push_back("TaskCaging");
+    _interfaces[PT_Module].push_back("VisualFeedback");
+}
+
+RManipulationPlugin::~RManipulationPlugin() {}
+
+OpenRAVE::InterfaceBasePtr RManipulationPlugin::CreateInterface(OpenRAVE::InterfaceType type, const std::string& interfacename, std::istream& sinput, OpenRAVE::EnvironmentBasePtr penv)
+{
+    switch(type) {
+    case PT_Module:
+        if( interfacename == "basemanipulation") {
+            return CreateBaseManipulation(penv);
+        }
+        else if( interfacename == "taskmanipulation" ) {
+            return CreateTaskManipulation(penv);
+        }
+        //else if( interfacename == "taskcaging") {
+        //    return CreateTaskCaging(penv);
+        //}
+        else if( interfacename == "visualfeedback") {
+            return CreateVisualFeedback(penv);
+        }
+        break;
+    default:
+        break;
+    }
+    return OpenRAVE::InterfaceBasePtr();
+}
+
+const RavePlugin::InterfaceMap& RManipulationPlugin::GetInterfaces() const
+{
+    return _interfaces;
+}
+
+const std::string& RManipulationPlugin::GetPluginName() const
+{
+    static std::string pluginname = "RManipulationPlugin";
+    return pluginname;
+}
 
 OPENRAVE_PLUGIN_API RavePlugin* CreatePlugin() {
     return new RManipulationPlugin();
